Restructure project, implement http_event_handler
- Restructures crates into subfolders (cli/, event_handlers/, functionality/, static/)
- Merge constants into separate crate (static/constants/)
- Merge types into separate crate (static/types/)
- Implement http_event_handler

+++ daemon/rust-impl/static/types/src/lib.rs
@@ -2,17 +2,17 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize)]
 pub struct Payload {
-	pub timestamp: f64,
-	pub endpoint: String,
+    pub timestamp: u128,
+    pub endpoint: String,
 }
 
 #[derive(Serialize, Deserialize)]
 pub struct Config {
-	pub endpoint: String,
+    pub endpoint: String,
 }
 
 #[derive(Serialize, Deserialize)]
 pub struct ValidationResult<'a> {
-	pub body: &'a str,
-	pub status: i32,
+    pub body: &'a str,
+    pub status: u16,
 }
